created two modes of operation of the program
Reworked the interactive mode of operation, wrote a non-interactive mode. I broke the code into modules.
