+++ rl.js
@@ -1,0 +1,7 @@
+const readline = require('readline');
+const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout
+});
+
+module.exports = rl;
